Backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Security and Cryptography
cryptography==41.0.7
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTTP and Networking
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Web Scraping and Browser Automation
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
lxml==4.9.3

# Natural Language Processing
spacy==3.7.2

# AI and Machine Learning
openai>=1.0.0

# Data Validation and Parsing
pydantic==2.5.0
email-validator==2.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0 